+++ familyhelper-ui-view/packages/component-api/src/api/clannad/profile.ts
@@ -1,9 +1,9 @@
 // noinspection JSUnusedGlobalSymbols
 
 import http from '../../util/http'
-import type { StringIdKey } from '../subgrade/key'
-import type { PagingInfo } from '../../util/request'
-import type { Pres } from '../../util/response'
+import { type StringIdKey } from '../subgrade/key'
+import { type PagingInfo } from '../../util/request'
+import { type Pres } from '../../util/response'
 import { ProfileIndicator } from './profileTypeIndicator'
 
 export type Profile = {