+++ familyhelper-ui-view/packages/component-api/src/api/note/noteItem.ts
@@ -1,12 +1,12 @@
 // noinspection JSUnusedGlobalSymbols
 
 import http from '../../util/http'
-import type { LongIdKey } from '../subgrade/key'
-import type { Pres, Prespa } from '../../util/response'
-import type { PagingInfo } from '../../util/request'
+import { type LongIdKey } from '../subgrade/key'
+import { type Pres, type Prespa } from '../../util/response'
+import { type PagingInfo } from '../../util/request'
 
-import type { DispNoteBook } from './noteBook'
-import type { DispNoteNode } from './noteNode'
+import { type DispNoteBook } from './noteBook'
+import { type DispNoteNode } from './noteNode'
 
 export type NoteItem = {
   key: LongIdKey