+++ familyhelper-ui-view/packages/component-api/src/api/note/noteNode.ts
@@ -1,11 +1,11 @@
 // noinspection JSUnusedGlobalSymbols
 
 import http from '../../util/http'
-import type { LongIdKey } from '../subgrade/key'
-import type { Pres, Prespa } from '../../util/response'
-import type { PagingInfo } from '../../util/request'
+import { type LongIdKey } from '../subgrade/key'
+import { type Pres, type Prespa } from '../../util/response'
+import { type PagingInfo } from '../../util/request'
 
-import type { DispNoteBook } from './noteBook'
+import { type DispNoteBook } from './noteBook'
 
 export type NoteNode = {
   key: LongIdKey