+++ familyhelper-ui-view/packages/component-api/src/api/note/ponb.ts
@@ -1,12 +1,12 @@
 // noinspection JSUnusedGlobalSymbols
 
 import http from '../../util/http'
-import type { LongIdKey } from '../subgrade/key'
-import type { Pres, Prespa } from '../../util/response'
-import type { PagingInfo } from '../../util/request'
+import { type LongIdKey } from '../subgrade/key'
+import { type Pres, type Prespa } from '../../util/response'
+import { type PagingInfo } from '../../util/request'
 
-import type { DispNoteBook } from './noteBook'
-import type { DispAccount } from '../system/account'
+import { type DispNoteBook } from './noteBook'
+import { type DispAccount } from '../system/account'
 
 export type PonbPermissionLevelOwner = 0
 export type PonbPermissionLevelGuest = 1