+++ familyhelper-ui-view/packages/component-api/src/api/notify/reset.ts
@@ -1,7 +1,7 @@
 // noinspection JSUnusedGlobalSymbols
 
 import http from '../../util/http'
-import type { Pres } from '../../util/response'
+import { type Pres } from '../../util/response'
 
 export function resetRoute(): Pres<null> {
   return http.generalClient().post('notify', 'reset-route/', {}, 'application/json;charset=UTF-8')