+++ familyhelper-ui-view/packages/component-api/src/api/settingrepo/reset.ts
@@ -1,7 +1,7 @@
 // noinspection JSUnusedGlobalSymbols
 
 import http from '../../util/http'
-import type { Pres } from '../../util/response'
+import { type Pres } from '../../util/response'
 
 export function resetFormat(): Pres<null> {
   return http