+++ familyhelper-ui-view/packages/component-api/src/api/system/login.ts
@@ -1,8 +1,8 @@
 // noinspection JSUnusedGlobalSymbols
 
 import http from '../../util/http'
-import type { LongIdKey, StringIdKey } from '../subgrade/key'
-import type { Pres } from '../../util/response'
+import { type LongIdKey, type StringIdKey } from '../subgrade/key'
+import { type Pres } from '../../util/response'
 
 export type DynamicLoginInfo = {
   account_key: StringIdKey