+++ familyhelper-ui-view/packages/component-api/src/api/system/time.ts
@@ -1,7 +1,7 @@
 // noinspection JSUnusedGlobalSymbols
 
 import http from '../../util/http'
-import type { ResponseData } from '../../util/response'
+import { type ResponseData } from '../../util/response'
 
 export function currentDate(): Promise<ResponseData<number>> {
   return http.publicClient().post('system', 'current-date', {}, 'application/json;charset=UTF-8')